appengine: Add support for scaling cover images.
Add a /cover handler to the server that reads the supplied
cover image from GCS or HTTP and writes a scaled JPEG
version of it.

I'm not sure if this is a good or bad idea. There's no good
reason for the Android client to be downloading full-sized
images, but from testing, this is slow enough that I turned
it off for the web client -- the /cover handler usually
takes half a second to run.