appengine/web: Update web client to use scaled cover images.
Update the web client to use the /cover handler to get
512x512 cover art images and use them everywhere apart from
the big background image in <presentation-layer>. Also
preload the covers for the next two tracks to get rid of
delays when switching.

I'm seeing the web client make four or so requests for the
same image, apparently because AppEngine rewrites the
Cache-Control and Expires HTTP headers in response to
requests from the admin user. That sucks, but it looks like
the requests after the first one usually get fulfilled from
memcache, so hopefully it's not a huge deal.