web: Check for errors before using preloaded song.
Make the music-player component check that the preloaded
audio element didn't report any errors before trying to use
it. When preloading fails due to a network error, the
element appears to get stuck in a state where it throws "The
element has no supported sources." exceptions even after the
network is back. Fixes #16.