update_music: Add support for specifying artist rewrites.
Add an 'artistRewrites' field to update_music's JSON config
that can be used to rewrite artist names from ID3 tags. I'm
adding this as a quick workaround for #18: if an artist name
gets updated in MusicBrainz, this provides a way to quickly
update the database without needing to retag and reupload
the music files.