update_music: Handle old files moved into music dir.
Make update_music track the directories that contained songs
in the previous run so it can identify new directories
(whose files may have stale timestamps).

The config file's lastUpdateTimeFile field is renamed to
lastUpdateInfoFile, and existing files need to be manually
converted to the new format.

I think that there are still edge cases (e.g. a directory
gets replaced by an identically-named directory with
different files that also have old timestamps), but that
seems esoteric enough that I'm not going to deal with it. I
mostly want to handle the case where I download an album but
forget to copy it under the music dir before running
update_music.

This fixes #22.