server: Kill off server/types module.
Untangle the circular dependency between the query and cache
modules by making cache contain generic caching code and
making query depend on it. Also make the update module call
into query to flush cached data.

Also move everything apart from Song and Play out of the
types module and move it to the top level of the repository.
It'd be nice to rename it, but the only name I can come up
with is "song" and I'd like to avoid song.Song redundancy.