server: Move types package to server/db.
I'd previously moved this package (which just defines the
Song type and friends) out of server/, but on further
reflection it seems mostly closely linked to the server,
since it defines both the Datastore representation and the
JSON representation sent and received by the server.

Also move the SongOrError type and the GetSongsFromChannel
test helper function into update_music, since that's the
only place where they're used.