web: Add song details dialog.
Add a new 'Details' item to the music-searcher and
music-player context menus to pop up a dialog containing
information about the song. This doesn't include playback
history, but maybe it should (that would require fetching it
from the server, though).

Also rename the dialog-manager component to overlay-manager
and make it handle Escape keypresses for all dialogs.