test: Port update-retrying web tests to Go.
Also update music-searcher to not focus the keyword input in
response to '/' while music-player's update div is shown.
The impetus for this is a ChromeDriver bug that results in a
'/' JS event being emitted when the 'z' key is sent on a
system that only has the Dvorak keyboard layout configured.