cmd/nup,server: Store server config in Datastore.
Add support for reading the server config from Datastore
instead of passing it via a deployed file, and add a new
'config' subcommand to the nup executable for printing and
setting the config.

Also delete the project_id.sh script and add a 'projectid'
subcommand that the deploy.sh script can use.