cmd/nup: Add -delete-instances to 'config' subcommand.
Add support for stopping all running instances to pick up
config changes. It'd be nicer if it were possible to just
send a request to all instances telling them to reload the
config from Datastore, but there's apparently no way to send
requests to specific instances when using automatic scaling.