server: Improve perceived randomness of shuffled songs.
Make /query?shuffle=1 reorder results to make it less likely
that songs from a single artist will be played back-to-back
or that songs from a single album will be played in sequence
for a given artist.

Unrelatedly, fix a bug where the server wasn't waiting for
cache writes to finish for queries with zero results.

Closes #28.