cmd/nup, server: Normalize slices in Song more robustly.
Add a Clean function to song that sorts and dedupes the
Keywords, Tags, and Plays fields, and make Update call it.
This should guard against adding duplicate plays if the same
songs are merged multiple times.

Also add a -delete-after-merge flag to 'nup update', and
preserve the highest rating when merging.