test: Don't automatically update datastore indexes.
Update test.DevAppserver to avoid automatically updating
index.yaml when a datastore query can't be satisfied with
the existing indexes, and improve the way that options are
passed.

Also update end-to-end tests to run a bunch of common
queries to make sure that the current indexes are
sufficient. Doing this resulted in a few more indexes being
needed. The test suite now accepts a -create-indexes flag to
indicate that dev_appserver.py should update index.yaml.

See #35. I'll try to clean up the indexes in a later change.