server: Serve last song from memory.
Make the server hold the last song requested via the /song
endpoint in memory if it was 32 MB or smaller.

Chrome usually sends multiple requests for each song (and
also drops the connection before reading all of the data
that it requested :-/), so this cuts down on unnecessary
reads from Cloud Storage.