server: Add /query fallback for missing indexes.
Add a slow fallback mode to the server's /query endpoint to
handle queries that initially fail due to not having
appropriate built-in or composite indexes. It's unfortunate
that datastore doesn't support doing this automatically.

I'm also deleting the autogenerated composite indexes in
index.yaml and just adding a few for presets that I use
frequently.

Tangentially, I special-cased the MaxPlays == 0 case to just
use an equality filter.

This mostly takes care of #35, but I might still experiment
with updating the UI to allow specifying multiple inequality
filters now that the backend can hopefully deal with it.