web: Minor improvements to tag suggestions.
Make the tag-suggester component provide suggestions when
the caret is positioned at the very beginning of a known
tag. This helps when you jump to the beginning of the text
field and start typing a tag.

Also fix some weird behavior that's been around since the
beginning of this repo's history where focusing the text
field would move the caret to the end of it. I think I did
this to override browsers' select-all-when-tabbing-to-field
behavior, but the right thing to do is to clear the
selection while preserving the caret position.

Also update example songs to have tags and different ratings
to make it easier to manually test the code.