cmd/nup: Rework mp3gain code for 'update' command.
Make the 'update' command avoid computing gain adjustments
for all of the files in each directory, and instead only
compute adjustments for files from the same album. The old
approach was very slow when a new file was added to a
directory containing a bunch of miscellaneous songs.

Also limit scanForUpdatedSongs to read 8 songs at once, as
it can run out of file descriptors otherwise when there are
a lot of updates.