web: Use <dialog> to display modal dialogs and menus.
Delete the overlay-manager component and add simple
dialog.js and menu.js files that export functions to open
modal dialogs and context menus using the dialog element.

I'm not completely sure that this is a good idea. Per
https://caniuse.com/dialog, Chrome has
supported the dialog element since 2014, but support was
just added to Firefox 98 and Safari 15.4 in March 2022. For
what it's worth, it seems to work reasonably well in
Chrome 100.0.4896.133. Pretty much anything is better than
the code that I'd written for overlay-manager, I figure.