Update Song.Rating to be int rather than float.
I think I initially made ratings be floating-point numbers
in the range [0.0, 1.0] (with -1.0 for unrated) for the sake
of "flexibility".

Then I went ahead and rated all my music as 0, 0.25, 0.5,
0.75, or 1 and wrote tons of code that converts these
ratings to stars in [1, 5]. To make things even worse,
SearchPreset.MinRating already uses stars. Using -1 to
represent unrated is also annoying since the default zero
value in Go is 0.

This change converts Song.Rating to be an int and adds a
custom datastore.PropertyLoadSaver implementation that maps
the old Rating and RatingAtLeast* properties to the new
ones. This code can be deleted after reindexing the existing
songs.