Support WebP cover images.
Make the server's /cover endpoint return a WebP version of
the image if a webp=1 parameter is supplied. There still
doesn't seem to be any way to generate WebP images from Go
on App Engine, so add a -generate-web flag to "nup covers"
that generates 256x256 and 512x512 versions of all of the
JPEGs in the cover directory.