web: Speculative fixes for audio-wrapper leaks.
Chrome heap snapshots imply that one "Detached
HTMLAudioElement" and one "Detached
MediaElementAudioSourceNode" are being retained for each
song that's been played.

I'm not sure if this is something that I'm doing wrong or a
Chrome bug, but try harder to detach source nodes, and
create new <audio> elements by cloning the template instead
of the existing <audio> element.