web: Fix preload URL safety check in audio-wrapper.
Fix preloading silently failing due to Chrome refusing to
load media. I suspect that this was caused by 09f867323f34.
Also log and throw away preload elements that have errors.

I unfortunately can't think of a good way to write a test
for this. I'd rather not add server code to track /song
requests so that a test can verify that the next song is
preloaded.