server, web: Use RFC 3339 and support date searches.
Update the server to accept RFC 3339 times in addition to
seconds since the Unix epoch, and update the web interface
and tests to use RFC 3339 times everywhere.

Also update the /query endpoint to accept minDate and
maxDate parameters that filter based on Song.Date for #42.
The web interface doesn't expose these yet.