server: Distinguish between normal and admin users.
Update the server's Config struct to contain a single list
of users instead of distinguishing between Google
authentication and HTTP basic authentication, and add the
ability to designate a user as being an administrator.

This makes it easier to restrict access to server endpoints
that are only called by the command-line executable. In a
future change, I plan to add support for "read-only" users
that can't rate or tag songs.