server: Append users' excluded tags in /query handler.
Android already drops excluded songs from its local
database, but it seems better to additionally not return
them in response to remote searches. This also affects the
songs returned to the web interface.

Also improve comments in the User struct and reject
configurations that include guest users with email
authentication.