web: Make Updater not persist state in-memory.
Change the Updater class to only use localStorage to store
queued/active plays, ratings, and tags instead of
additionally storing them in instance variables. This isn't
intended to make any major changes to the class's logic, but
it'll hopefully make it easier to avoid races if multiple
tabs are open in a later change. See #51.