web: Make Updater wait before sending when coming online.
If the web interface is open is multiple tabs, bad things
are likely to happen if all of the Updater instances start
stealing each other's queued records from localStorage when
an 'online' event is received. To (hopefully) prevent this
from happening, make each instance sleep a different amount
of time before sending. Fixes #51.