web: Only examine prev/next songs for auto gain type.
Update play-view to use album-specific gain adjustments for
the "auto" gain type if the previous or next song is from
the same album as the current song, and track-specific
adjustments otherwise. See #54.

The previous approach that examined the whole playlist
handled some situations poorly, e.g. when a full album was
appended to the end of a shuffled playlist, or an "unrated"
query returned a bunch of full albums but also a singleton
from an otherwise-entirely-rated album.