Extract and store disc subtitles.
Add a new DiscSubtitle field to db.Song and update the nup
command-line tool to set it from the TSST (Set subtitle)
frame or from the end of rewritten album names (e.g.
" (disc 2: Subtitle)"). The server includes the subtitle in
the song's keywords, and the web interface's
song-info-dialog component includes it in a new "Disc" row.

This also updates 5s.mp3 to have a subtitle at the end of
its TALB frame and 10s.mp3 to have a new TSST frame
containing a subtitle.