cmd/nup: Fix metadata override update bug.
Fix a dumb-but-subtle bug in the 'metadata' subcommand. The
updated override file was computed by diffing against a
version of the song which had the existing override file
applied to it, so previous changes could be undone in the
new version of the override file.

Add a new files.UpdateMetadataOverride function that rereads
the song's ID3 metadata and uses it for the diff instead.